replaced original challenge and added a second
The original challenge contained too many new concepts that will confuse novice users (e.g. adding directories to staging and ignoring directories).  I removed these concepts as they are not needed at this stage.  The main point is that you don't need to create a git repo for each sub-directory, the top-level report will track everything under it.  The user also does not need to know the intricate implications of nested repos.  As per the discussion in issue /issues/272 (=/=swcarpentry/=/=git-novice/issues/272), this is an advanced topic, and advanced users may have use for them and be able to manage them.  Novice users only n need to understand it creates confusion and is redundant, so as a general rule it is not a good idea.  I wanted to keep the last bit about deleting the .git folder.  While this is not recommended, it reinforces the concept that it is the .git folder that makes an ordinary directory a git repository.  Therefore, if you accidentally run init in the wrong directory you can easily correct your error by removing this folder.