Update 09-conflict.md
These changes clarify when a merge conflict is identified by git. In this case, the git push command fails not because of the merge conflict, but because the remote repository has newer changes that have not yet been incorporated into the local repository. Even if a merge conflict does not exist, git pull will fail in this scenario. The merge conflict is only identified after git fetch retrieves the changes made to the remote repository into the local version and then attempts to merge the two versions with git merge.  This necessitates manual intervention to complete the failed merge operation.

These changes are part of my software carpentry instructor training checkout process.