Add explanations for additional git log flags
Adds explanations for the flags used in the example for how to combine multiple
git log flags with `--oneline` and corrects the output for the first `git log
--oneline` example where the commit hashes should not have "*" prefixes (these
prefixes are drawn by the `--graph` flag).

For clarity, this commit also removes the `--all` flag from the combined flags
example since this flag does not change the output of git log in the simple
repository of this lesson and requires a longer explanation of git references
than there is time to cover during the novice lesson. The description of the
`--decorate` flag does provide a link to the official documentation for git
references, for learners who wish to study this subject in more detail on their
own.

Closes /issues/536.