change usage of "lesson" in episode 2
Neither usage of "lesson" appeared correct, according to Carpentries nomenclature.

The first instance of "lesson" seemed to contradict the second - one indicated that we may push to GitLab or BitBucket later in the lesson, the other that we would stick with GitHub. I changed this to indicate what I think was the important message here: that the name and email supplied here will have consequences beyond the lesson the learner is working on presently.

The second was a simpler change; "these lessons" became "this lesson"; I'm not aware of any other Carpentries Git lessons that this sentence could ahve been referring to.